prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -5,8 +5,8 @@
  */
 int main(void)
 {
-	long long int num = 612852475143;
-	int n;
+	unsigned long int num = 612852475143;
+	long unsigned int n;
 
 	n = 2;
 	while (num != 1)
